Deletes unexisting raise #7
