+++ app/controllers/api/v1/matches/confirmations_controller.rb
@@ -1,7 +1,9 @@
 module Api
   module V1
     module Matches
-      class ConfirmationsController < BaseController
+      class ConfirmationsController < ApiController
+        include Api::V1::MatchesConcern
+
         before_action :load_match, only: [:update]
         before_action :check_auth_admin, only: [:update]
