+++ app/controllers/api/v1/matches/finishes_controller.rb
@@ -1,14 +1,14 @@
 module Api
   module V1
     module Matches
-      class FinishesController < BaseController
+      class FinishesController < ApiController
+        include Api::V1::MatchesConcern
+
         before_action :load_match, only: [:update]
         before_action :check_auth_admin, only: [:update]
 
         def update
           @match.finish
-        rescue ActiveRecord::RecordInvalid => error
-          error_request(message: error.message)
         end
       end
     end
